annotator modules added by passing config val (#90)
* annotator modules added by passing config val

* fix merge conflict

* following same pattern as parsers , modify configs

* fix to dug config method

* fix old dug pipeline for backward compatiblity

* correct default annotator type

* reflective changes

* typo extra quotes

* annotator type not being picked up from config

* remove annotate simple , log env value for lakefs enabled

* testing lakefs off

* add more logging

* add more logging

* post init for config to parse to boolean

* put back task calls

* revert some changes

* adding new pipeline

* lakefs io support for merge task

* fix name

* add io params for kg tasks

* wire up i/o paths for merge

* fix variable name

* print files

* few debug logs

* few debug logs

* treat path as path not str

* few debug logs

* some fixes

* logging edge files

* bug fix knowledge has edge

* re-org graph structure

* adding pathing for other tasks

* pagenation logic fix for avalon

* update lakefs client code

* fix glob for get kgx files

* fix up get merged objects

* send down fake commit id for metadata

* working on edges schema

* bulk create nodes I/O

* find schema file

* bulk create edges  I/O

* bulk create edges  I/O

* bulk load io

* no outputs for final tasks

* add recursive glob

* fix globbing

* oops

* delete dags

* pin dug to latest release

* cruft cleanup

* re-org kgx config

* add support for multiple initial repos

* fix comma

* create dir to download to

* swap branch and repo

* clean up dirs

* fix up other pipeline 👌

---------

+++ dags/roger/tasks.py
@@ -18,7 +18,7 @@ from roger.config import config, RogerConfig
 from roger.logger import get_logger
 from roger.pipelines.base import DugPipeline
 from avalon.mainoperations import put_files, LakeFsWrapper, get_files
-import lakefs_client
+from lakefs_sdk.configuration import Configuration
 from functools import partial
 
 logger = get_logger()
@@ -48,7 +48,7 @@ def task_wrapper(python_callable, **kwargs):
         output_data_path = generate_dir_name_from_task_instance(kwargs['ti'],
                                                             roger_config=config,
                                                             suffix='output')
-    else: 
+    else:
         input_data_path, output_data_path = None, None
     # cast it to a path object
     func_args = {
@@ -102,7 +102,7 @@ def get_executor_config(data_path='/opt/airflow/share/data'):
     return k8s_executor_config
 
 def init_lakefs_client(config: RogerConfig) -> LakeFsWrapper:
-    configuration = lakefs_client.Configuration()
+    configuration = Configuration()
     configuration.username = config.lakefs_config.access_key_id
     configuration.password = config.lakefs_config.secret_access_key
     configuration.host = config.lakefs_config.host
@@ -151,7 +151,7 @@ def avalon_commit_callback(context: DagContext, **kwargs):
     # right now lets stick to using one repo ,
 
     # issue Vladmir pointed out if uploads to a single lakefs branch have not
-    # been finilized with commit,
+    # been finalized with commit,
     # this would cause dirty commits if parallel tasks target the same branch.
 
     # solution: Lakefs team suggested we commit to a different temp branch per
@@ -173,7 +173,9 @@ def avalon_commit_callback(context: DagContext, **kwargs):
         s3storage=False,
         lake_fs_client=client,
         branch=temp_branch_name,
-        repo=repo
+        repo=repo,
+        # @TODO figure out how to pass real commit id here
+        commit_id=branch
     )
 
     # see what changes are going to be pushed from this branch to main branch
@@ -201,11 +203,20 @@ def avalon_commit_callback(context: DagContext, **kwargs):
         logger.info(f"deleted temp branch {temp_branch_name}")
         logger.info(f"deleting local dir {local_path}")
         files_to_clean = glob.glob(local_path + '**', recursive=True) + [local_path]
+    clean_up(context, **kwargs)
+
+def clean_up(context: DagContext, **kwargs):
+    input_dir = str(generate_dir_name_from_task_instance(context['ti'],
+                                                          roger_config=config,
+                                                          suffix='output')).rstrip('/') + '/'
+    output_dir = str(generate_dir_name_from_task_instance(context['ti'],
+                                                          roger_config=config,
+                                                          suffix='input')).rstrip('/') + '/'
+    files_to_clean = glob.glob(input_dir + '**', recursive=True) + [input_dir]
+    files_to_clean += glob.glob(output_dir + '**', recursive=True) + [output_dir]
     for f in files_to_clean:
-        shutil.rmtree(f)
-
-
-
+        if os.path.exists(f):
+            shutil.rmtree(f)
 
 def generate_dir_name_from_task_instance(task_instance: TaskInstance,
                                          roger_config: RogerConfig, suffix:str):
@@ -222,12 +233,12 @@ def generate_dir_name_from_task_instance(task_instance: TaskInstance,
         f"{root_data_dir}/{dag_id}_{task_id}_{run_id}_{try_number}_{suffix}")
 
 def setup_input_data(context, exec_conf):
-    print("""
+    logger.info("""
         - Figures out the task name and id,
         - find its data dependencies
         - clean up and create in and out dir
         - put dependency data in input dir
-        - if for some reason data was not found raise an execption
+        - if for some reason data was not found raise an exception
           """)
     # Serves as a location where files the task will work on are placed.
     # computed as ROGER_DATA_DIR + /current task instance name_input_dir
@@ -241,38 +252,48 @@ def setup_input_data(context, exec_conf):
 
     # Download files from lakefs and store them in this new input_path
     client = init_lakefs_client(config=config)
-    input_repo = exec_conf['input_repo']
-    input_branch = exec_conf['input_branch']
-    # If input repo is provided use that as source of files
-    if exec_conf.get('path') and exec_conf.get('path') == '*':
-        remote_paths =  ['*'] # root path to get all sub dirs
-    # else figure out what to pull from the repo based on task name etc...
-    else:
+    repos = exec_conf['repos']
+    # if no external repo is provided we assume to get the upstream task dataset.
+    if not repos or len(repos) == 0:
+        # merge destination branch
+        branch = config.lakefs_config.branch
+        repo = config.lakefs_config.repo
         task_instance: TaskInstance = context['ti']
         # get upstream ids
         upstream_ids = task_instance.task.upstream_task_ids
         dag_id = task_instance.dag_id
         # calculate remote dirs using dag_id + upstreams
+        repos = [{
+            'repo': repo,
+            'branch': branch,
+            'path': f'{dag_id}/{upstream_id}'
+        } for upstream_id in upstream_ids]
 
-        remote_paths = [f'{dag_id}/{upstream_id}'
-                        for upstream_id in upstream_ids]
-    for remote_path in remote_paths:
+    # input_repo = exec_conf['input_repo']
+    # input_branch = exec_conf['input_branch']
+    # If input repo is provided use that as source of files
+    for repo in repos:
+        if not repo.get('path'):
+            # get all if path is not specified
+            repo['path'] = '*'
+    logger.info(f"repos : {repos}")
+    for r in repos:
         logger.info("downloading %s from %s@%s to %s",
-                    remote_path, input_repo, input_branch, input_dir)
+                    r['path'], r['repo'], r['branch'], input_dir)
+        # create path to download to ...
+        if not os.path.exists(input_dir + f'/{r["repo"]}'):
+            os.mkdir(input_dir + f'/{r["repo"]}')
         get_files(
-            local_path=input_dir,
-            remote_path=remote_path,
-            branch=input_branch,
-            task_name=context['task'].task_id,
-            pipeline_id=context['dag'].dag_id,
-            repo=input_repo,
+            local_path=input_dir + f'/{r["repo"]}',
+            remote_path=r['path'],
+            branch=r['branch'],
+            repo=r['repo'],
             changes_only=False,
-            metafilename=None,
             lake_fs_client=client
         )
 
-def create_python_task(dag, name, a_callable, func_kwargs=None, input_repo=None,
-                       input_branch=None, pass_conf=True, no_output_files=False):
+
+def create_python_task(dag, name, a_callable, func_kwargs=None, external_repos = {}, pass_conf=True, no_output_files=False):
     """ Create a python task.
     :param func_kwargs: additional arguments for callable.
     :param dag: dag to add task to.
@@ -305,25 +326,23 @@ def create_python_task(dag, name, a_callable, func_kwargs=None, input_repo=None,
 
         # repo and branch for pre-execution , to download input objects
         pre_exec_conf = {
-            'input_repo': config.lakefs_config.repo,
-            'input_branch': config.lakefs_config.branch
+            'repos': []
         }
-
-        if input_repo and input_branch:
-            # if the task is a root task , begining of the dag...
+        if external_repos:
+            # if the task is a root task , beginning of the dag...
             # and we want to pull data from a different repo.
             pre_exec_conf = {
-                'input_repo': input_repo,
-                'input_branch': input_branch, 
-                # if path is not defined , we can use the context (dag context) to
-                # resolve the previous task dir in lakefs.
-                'path': '*'
+                'repos': [{
+                    'repo': r['name'],
+                    'branch': r['branch'],
+                    'path': r.get('path', '*')
+                } for r in external_repos]
             }
             
         pre_exec = partial(setup_input_data, exec_conf=pre_exec_conf)
         # add pre_exec partial function as an argument to python executor conf 
         python_operator_args['pre_execute'] = pre_exec
-
+        python_operator_args['on_failure_callback'] = partial(clean_up, kwargs=op_kwargs)
         # if the task has  output files, we will add a commit callback  
         if not no_output_files:
             python_operator_args['on_success_callback'] = partial(avalon_commit_callback, kwargs=op_kwargs)
@@ -352,8 +371,10 @@ def create_pipeline_taskgroup(
                 dag,
                 f"annotate_{name}_files",
                 pipeline.annotate,
-                input_repo=getattr(pipeline_class, 'pipeline_name'),
-                input_branch=input_dataset_version,
+                external_repos=[{
+                    'name': getattr(pipeline_class, 'pipeline_name'),
+                    'branch': input_dataset_version
+                }],
                 pass_conf=False)
 
             index_variables_task = create_python_task(
